+++ plugins/decl_hdf5/tests/decl_hdf5_test_01.c
@@ -1,5 +1,6 @@
 /*******************************************************************************
  * Copyright (C) 2015-2019 Commissariat a l'energie atomique et aux energies alternatives (CEA)
+ * Copyright (C) 2020 Institute of Bioorganic Chemistry Polish Academy of Science (PSNC)
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -38,7 +39,7 @@ const char* CONFIG_YAML =
     "  test_var: double            \n"
     "plugins:                      \n"
     "  decl_hdf5:                  \n"
-    "    file: $meta1.h5           \n"
+    "    file: ${meta1}.h5         \n"
     "    write: [ test_var, meta2 ]\n"
     ;